+++ src/Products/PloneMeeting/config.py
@@ -76,8 +76,7 @@ setDefaultRoles(ManageOwnOrganizationFields, ())
 # and .../meeting-config-id/classifiers folder
 setDefaultRoles(ManageItemCategoryFields, ())
 setDefaultRoles(WriteUseridField, ('Manager', ))
-# the 'meeting_remove_meetingobserverglobal' WFA is used in various code
-MEETING_REMOVE_MOG_WFA = 'meeting_remove_meetingobserverglobal'
+MEETING_REMOVE_MOG_WFA = 'meeting_remove_global_access'
 
 # Permissions
 DEFAULT_ADD_CONTENT_PERMISSION = "Add portal content"